Better and safer way to access nconf
