+++ server/src/middleware/validatekey.js
@@ -1,4 +1,4 @@
-import nconf from 'nconf';
+import { nconf } from '../config.js';
 import { getLogger } from '@sitespeed.io/log';
 const logger = getLogger('sitespeedio.server');
 import { getText } from '../util/text.js';
