+++ server/src/middleware/validatescripting.js
@@ -1,6 +1,6 @@
 import validator from 'validator';
 import { getLogger } from '@sitespeed.io/log';
-import nconf from 'nconf';
+import { nconf } from '../config.js';
 
 const { isURL } = validator;
 const logger = getLogger('sitespeedio.server');
