+++ server/src/routes/html/admin/index.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 
-import nconf from 'nconf';
+import { nconf } from '../../../config.js';
 import { getText } from '../../../util/text.js';
 import {
   getExistingQueue,
