+++ server/src/util/text.js
@@ -1,4 +1,4 @@
-import nconf from 'nconf';
+import { nconf } from '../config.js';
 import y18n from 'y18n';
 import { getBaseFilePath } from './fileutil.js';
 const locale = nconf.get('localization:defaultLocale');
